combine update and write relation
